LayoutTests/inspector/runtime/getProperties-expected.txt


== Running test suite: Runtime.getProperties
-- Running test case: CheckPropertiesOfWrapperObject
Evaluating expression: (function(){var r = Object(5); r.foo = 'cat';return r;})()
Properties:
  __proto__ object Number
  foo string cat

-- Running test case: CheckPropertiesOfArray
Evaluating expression: ['red', 'green', 'blue']
Properties:
  __proto__ object Array
  0 string red
  1 string green
  2 string blue

-- Running test case: CheckPropertiesOfBoundConstructor
Evaluating expression: Number.bind({}, 5)
Properties:
  __proto__ function function () {
    [native code]
}
  arguments getter setter
  caller getter setter
  length number 0
  name string bound Number
Internal properties:
  boundArgs object Array
  boundThis object Object
  targetFunction function function Number() {
    [native code]
}

-- Running test case: CheckPropertiesOfBoundFunctionNoArguments
Evaluating expression: (function(){}).bind(null)
Properties:
  __proto__ function function () {
    [native code]
}
  arguments getter setter
  caller getter setter
  length number 0
  name string bound 
Internal properties:
  boundThis object undefined
  targetFunction function function (){}
